requirements.txt
torch>=2.0.0
transformers>=4.30.0
Pillow>=10.0.0
requests>=2.31.0
numpy>=1.24.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
replicate>=0.15.0
scikit-learn>=1.3.0
scipy>=1.11.0

openai>=1.0.0
